# Faster `B[:,i]` for a `BandedMatrix` using broadcasting