Currently, broadcasting is faster than `setindex!` as SIMD gets blocked in the latter  (https://github.com/JuliaLang/julia/issues/40962#issuecomment-1921340377), so this change uses the faster variant. This might need revisiting in the future if `setindex!` becomes significantly faster.

On master
```julia
julia> B = brand(6000,6000,4000,4000);

julia> @btime $B[:, 3];
  4.985 μs (4 allocations: 47.00 KiB) # master
  3.980 μs (4 allocations: 47.00 KiB) # PR

julia> @btime $B[:, 5000];
  5.619 μs (4 allocations: 47.00 KiB) # master
  4.462 μs (4 allocations: 47.00 KiB) # PR
```